+++ R/prop_allocation.R
@@ -1,10 +1,10 @@
 #---- Internal helpers ----
-# Rounding methods
+# Apportionment (rounding) methods
 largest_remainder <- function(p, n, initial) {
-  np <- p * (n / sum(p))
-  npf <- floor(np)
-  res <- initial + npf
-  remainder <- order(npf - np)[seq_len(n - sum(npf))]
+  p <- p * (n / sum(p))
+  np <- floor(p)
+  res <- initial + np
+  remainder <- order(np - p)[seq_len(n - sum(np))]
   res[remainder] <- res[remainder] + 1
   res
 }
@@ -22,9 +22,6 @@ highest_averages <- function(divisor) {
   )
   # return function
   function(p, n, initial) {
-    if (is.null(names(initial))) {
-      names(initial) <- names(p)
-    }
     while (n > 0) {
       i <- which.max(p / divisor(initial))
       initial[i] <- initial[i] + 1
@@ -34,14 +31,13 @@ highest_averages <- function(divisor) {
   }
 }
 
-#---- Expected coverage ----
-expected_coverage <- function(x, N, s = gl(1, length(x))) {
+# Argument checking
+check_allocation <- function(x, N, s) {
   if (not_strict_positive_vector(x)) {
     stop(
       gettext("'x' must be a strictly positive and finite numeric vector")
     )
   }
-  N <- trunc(N)
   if (not_positive_number(N)) {
     stop(
       gettext("'N' must be a positive and finite number")
@@ -52,7 +48,6 @@ expected_coverage <- function(x, N, s = gl(1, length(x))) {
       gettext("sample size 'N' is greater than population size")
     )
   }
-  s <- as.factor(s)
   if (length(x) != length(s)) {
     stop(
       gettext("'x' and 's' must be the same length")
@@ -63,79 +58,67 @@ expected_coverage <- function(x, N, s = gl(1, length(x))) {
       gettext("'s' cannot contain NAs")
     )
   }
+}
+
+#---- Expected coverage ----
+expected_coverage <- function(x, N, s = gl(1, length(x))) {
+  N <- trunc(N)
+  s <- as.factor(s)
+  check_allocation(x, N, s)
   p <- split(log(1 - .inclusion_prob(x, N)), s)
   sum(1 - vapply(p, function(x) exp(sum(x)), numeric(1L)))
 }
 
 #---- Proportional allocation ----
 prop_allocation <- function(
-    x, N, s = gl(1, length(x)), min = 0, 
+    x, N, s = gl(1, length(x)), initial = 0, 
     method = c("Largest-remainder", "D'Hondt", "Webster", "Imperiali", 
                "Huntington-Hill", "Danish", "Adams", "Dean")  
 ) {
-  if (not_strict_positive_vector(x)) {
-    stop(
-      gettext("'x' must be a strictly positive and finite numeric vector")
-    )
-  }
   N <- trunc(N)
-  if (not_positive_number(N)) {
-    stop(
-      gettext("'N' must be a positive and finite number")
-    )
-  }
-  min <- trunc(min)
-  if (not_positive_number(min)) {
-    stop(
-      gettext("'min' must be a positive and finite number")
-    )
-  }
-  if (N > length(x)) {
-    stop(
-      gettext("sample size 'N' is greater than population size")
-    )
-  }
   s <- as.factor(s)
-  if (length(x) != length(s)) {
+  check_allocation(x, N, s)
+  initial <- trunc(initial)
+  if (not_positive_vector(initial)) {
     stop(
-      gettext("'x' and 's' must be the same length")
+      gettext("'initial' must be a positive and finite numeric vector")
     )
   }
-  if (anyNA(s)) {
+  if (length(initial) == 1L) initial <- rep.int(initial, nlevels(s))
+  if (length(initial) != nlevels(s)) {
     stop(
-      gettext("'s' cannot contain NAs")
+      gettext("'initial' must have a single allocation size for each level in 's'")
+    )
+  }
+  ns <- tabulate(s, nbins = nlevels(s))
+  if (any(initial > ns)) {
+    stop(
+      gettext("'initial' must be smaller than the population size for each stratum")
+    )
+  }
+  N <- N - sum(initial)
+  if (N < 0) {
+    stop(
+      gettext("initial allocation is larger than 'N'")
     )
   }
   method <- match.arg(method)
-  ns <- tabulate(s, nbins = nlevels(s))
-  if (any(min > ns)) {
-    stop(
-      gettext("'min' must be smaller than the population size for each stratum")
-    )
-  }
-  N <- N - min * nlevels(s)
-  if (N < 0) {
-    stop(
-      gettext("minimal allocation is larger than 'N'")
-    )
-  }
-  # apportionment method
-  round <- if (method == "Largest-remainder") {
+  apportionment <- if (method == "Largest-remainder") {
     largest_remainder
   } else {
     highest_averages(method)
   }
+  res <- initial
   p <- vapply(split(x, s), sum, numeric(1L))
-  res <- structure(rep(min, length(p)), names = levels(s))
-  # redistribute sample units for those that cap out at the stratum size
   repeat {
-    res <- round(p, N, res)
+    res <- apportionment(p, N, res)
     d <- pmax(res - ns, 0)
     over <- which(as.logical(d))
     if (length(over) == 0L) break
+    # redistribute sample units for those that cap out at the stratum size
     res[over] <- ns[over]
     p[over] <- 0  
     N <- sum(d)
   }
-  res
+  structure(res, names = levels(s))
 }