+++ NEWS.md
@@ -2,7 +2,7 @@
 
 - `prop_allocation()` gains two new arguments:
 
-    1. `min` sets the minimum allocation for each stratum. This saves from having to manually keep track of an initial allocation and adjust the sample size passed to `prop_allocation()`.
+    1. `initial` sets the initial allocation for each stratum. This saves from having to manually keep track of an initial allocation and adjust the sample size passed to `prop_allocation()`.
 
     2. `method` selects the apportionment method used to round a proportional allocation to integer values. In particular, highest-averages methods can be used in place of the largest-remainder method.